requirements.txt

streamlit
openai
chromadb
PyGithub
tiktoken
python-dotenv
